Bug fixed. use BigDecimal replaced Double.

+++ app/src/main/java/com/example/calculator/Sirusi.java
@@ -1,9 +1,11 @@
 package com.example.calculator;
 
+import java.math.BigDecimal;
+
 public class Sirusi {
     //これは僕の生きる印
-    public static Double operand1=null;
+    public static BigDecimal operand1=null;
     public static String operate="null";
-    public static Double operand2=null;
+    public static BigDecimal operand2=null;
 
 }
